requirements.txt
requests
python-dotenv
moviepy
gTTS
